% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_check.R
\name{check_github_suite_ver}
\alias{check_github_suite_ver}
\title{Check for updates to Giotto Suite}
\usage{
check_github_suite_ver(pkg = "Giotto")
}
\arguments{
\item{pkg}{character. Package to check (pattern matches)}
}
\description{
Checks the Giotto Suite github repository and compares the version
number to the currently installed.
}
\examples{
check_github_suite_ver("GiottoUtils")
}
\keyword{internal}
