% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{color_tag}
\alias{color_tag}
\alias{color_red}
\alias{color_green}
\alias{color_yellow}
\alias{color_blue}
\alias{color_purple}
\alias{color_teal}
\alias{use_color_text}
\alias{ansi_colors}
\alias{is_emacs_with_color}
\alias{emacs_version}
\title{Colorize print text}
\usage{
color_tag()

color_red(x)

color_green(x)

color_yellow(x)

color_blue(x)

color_purple(x)

color_teal(x)

use_color_text()

ansi_colors()

is_emacs_with_color()

emacs_version()
}
\value{
named list of characters
}
\description{
Return a list of 8 formatting tags for ansi colored text
}
\details{
supported colors checking is modified from \pkg{cli}
\href{https://github.com/r-lib/cli/blob/HEAD/R/num-ansi-colors.R}{aab-num-ansi-colors.R}
}
\section{Functions}{
\itemize{
\item \code{use_color_text()}: Determine if system should print color

\item \code{ansi_colors()}: Determine if system can print at least 8 colors

\item \code{is_emacs_with_color()}: Determine if emacs can print color

\item \code{emacs_version()}: Determine emacs version

}}
\examples{
color_tag()

color_red()

color_green()

color_yellow()

color_blue()

color_purple()

color_teal()

use_color_text()

ansi_colors()

is_emacs_with_color()

emacs_version()

}
\keyword{internal}
