% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_helpers.R
\name{dcast_dt_string}
\alias{dcast_dt_string}
\title{dcast_dt_string}
\usage{
dcast_dt_string(data, col_name1, col_name2, value.var)
}
\arguments{
\item{data}{a \code{data.table}}

\item{col_name1}{character. LHS of cast formula}

\item{col_name2}{character. RHS of cast formula}

\item{value.var}{character. Name of the column whose values will be filled to
cast.}
}
\description{
Data.table dcast using character inputs for formula
}
\examples{
library(data.table)
ChickWeight = as.data.table(ChickWeight)
setnames(ChickWeight, tolower(names(ChickWeight)))
DT <- melt(as.data.table(ChickWeight), id=2:4) # calls melt.data.table

dcast_dt_string(DT, 'chick', 'time', 'value')
}
\seealso{
\code{\link[data.table:dcast.data.table]{data.table::dcast.data.table()}}
}
\concept{data.table}
