% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_info.R
\name{determine_cores}
\alias{determine_cores}
\title{determine_cores}
\usage{
determine_cores(
  cores = getOption("giotto.cores", default = NA),
  min_cores = 1,
  max_cores = 10
)
}
\value{
numeric
}
\description{
guesses how many cores to use
}
\examples{
determine_cores()
}
\keyword{internal}
