% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.log_create}
\alias{.log_create}
\title{Create a log file}
\usage{
.log_create(filedir = getOption("giotto.logdir", tempdir()))
}
\arguments{
\item{filedir}{character. Directory to create a logfile}
}
\value{
file 'log.txt'
}
\description{
Creates a file called 'log.txt' at the specified directory. If no directory
is provided, it defaults to \code{tempdir()}, but a specific one can be provided
by setting it to the "giotto.logdir" option or using \code{.log_dir()}.
The filepath is additionally written to the option 'giotto.last_logpath'.
}
\keyword{internal}
