% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.log_dir}
\alias{.log_dir}
\title{Giotto logging directory}
\usage{
.log_dir(logdir = tempdir())
}
\arguments{
\item{logdir}{character. Directory to log to (defaults to tempdir)}
}
\description{
Set the Giotto logging directory. New logfiles will be generated here.
}
\keyword{internal}
