% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_helpers.R
\name{dt_set_row_order}
\alias{dt_set_row_order}
\title{Set specific data.table row order}
\usage{
dt_set_row_order(x, neworder)
}
\arguments{
\item{x}{data.table}

\item{neworder}{numerical vector to reorder rows}
}
\description{
Set specific data.table row order
}
\concept{data.table}
