% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_helpers.R
\name{dt_sort_combine_two_columns}
\alias{dt_sort_combine_two_columns}
\title{dt_sort_combine_two_columns}
\usage{
dt_sort_combine_two_columns(DT, column1, column2, myname = "unif_gene_gene")
}
\arguments{
\item{DT}{datatable}

\item{column1}{column1 to use}

\item{column2}{column2 to use}

\item{myname}{name of combined column to generate}
}
\description{
fast sorting and pasting of 2 character columns in a data.table
}
\concept{data.table}
