% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{file_extension}
\alias{file_extension}
\title{Get file extension}
\usage{
file_extension(file)
}
\arguments{
\item{file}{character. Filepath}
}
\value{
character
}
\description{
Get the file extension from a provided filepath
}
\examples{
file_extension("my_file.txt")

}
