% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{file_extention}
\alias{file_extention}
\alias{file_extension}
\title{Get file extention}
\usage{
file_extension(file)
}
\arguments{
\item{file}{character. Filepath}
}
\description{
Get the file extention from a provided filepath
}
