% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{fread_colmatch}
\alias{fread_colmatch}
\title{Fread specific rows based on column matches}
\usage{
fread_colmatch(file, col, sep = NULL, values_to_match, verbose = FALSE, ...)
}
\arguments{
\item{file}{path to file to load}

\item{col}{name of col to match from}

\item{sep}{grep term to match as column delimiters within the file}

\item{values_to_match}{values in \code{col} to match given as a vector}

\item{verbose}{whether to print the grep command}

\item{...}{additional parameters to pass to \code{\link[data.table]{fread}}}
}
\value{
A data.table
}
\description{
Fread specific rows based on column matches
}
\examples{
x = data.frame(a = c("a","b","c"), b = 1:3, c = 5:7)
write.csv(x, "my_file.csv")
fread_colmatch("my_file.csv", col = "a", values_to_match = c(1,3))

}
\keyword{internal}
