% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmatch.R
\name{g_match_arg}
\alias{g_match_arg}
\title{Partial matching of character choices}
\usage{
g_match_arg(arg, choices, ...)
}
\arguments{
\item{arg}{character. Argument to match}

\item{choices}{character vector of choices to match to}

\item{\dots}{additional params passed to grep}
}
\description{
Given an arg to match to, tries to partially match to the provided characeter
vector of choices. Matching ignores case and returns the first match.
}
\keyword{internal}
