% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{getMonochromeColors}
\alias{getMonochromeColors}
\title{Create color scaling for a single color starting from black}
\usage{
getMonochromeColors(col, n = 256L, ...)
}
\arguments{
\item{col}{hexadecimal color to scale scale towards}

\item{n}{number of colors to request in monochrome palette}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette}}
  \describe{
    \item{\code{bias}}{a positive number.  Higher values give more widely spaced
    colors at the high end.}
    \item{\code{space}}{a character string; interpolation in RGB or \abbr{CIE} Lab
    color spaces.}
    \item{\code{interpolate}}{use spline or linear interpolation.}
    \item{\code{alpha}}{logical: should alpha channel (opacity) values be
    returned?   It is an error to give a true value if
    \code{space} is specified.}
  }}
}
\value{
character vector
}
\description{
Create color scaling for a single color starting from black
}
\examples{
getMonochromeColors("green", n = 100)
}
\seealso{
Other basic color palette functions:
\code{\link[=getDistinctColors]{getDistinctColors()}},
\code{\link[=getRainbowColors]{getRainbowColors()}}
}
\concept{basic color palette functions}
