% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{gstop}
\alias{gstop}
\title{Module-specific error message}
\usage{
gstop(
  ...,
  sep = " ",
  strWidth = 100,
  errWidth = FALSE,
  .module,
  .prefix = " ",
  .initial = "",
  .n = 1L,
  .call = TRUE,
  .warn_nstack = getOption("giotto.warn_gstop_nstack", FALSE)
)
}
\arguments{
\item{...}{additional strings and/or elements to pass to wrap_msg}

\item{sep}{how to join elements of string (default is one space)}

\item{strWidth}{externally set wrapping width.
(default value of 100 is not effected)}

\item{errWidth}{default = FALSE. Set strWidth to be
compatible with error printout}

\item{.module}{character. Giotto module to send the error from}

\item{.prefix}{character. prefix for later lines}

\item{.initial}{character. prefix for first line}

\item{.n}{stack frames back where the error happened}

\item{.call}{logical, whether to include the call selected through .n as the
location where the error was. Default is TRUE}

\item{.warn_nstack}{logical. whether to warn when there are insufficient
stackframes for requested .n (default = FALSE)}
}
\value{
character message
}
\description{
Send an error message formatted with \code{wrap_txt()}. Prepends the Giotto
module from which the error was triggered. This function should not be used
directly, but instead an internal \code{.gstop()} function should be created using
this framework for each module using it. The \code{.n} param should be incremented
to 2L for this wrapper function
}
\examples{
try(
    gstop("My stop message", .module = "GiottoUtils"),
    silent = TRUE
)
}
