% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_handle.R
\name{handle_warnings}
\alias{handle_warnings}
\title{Warning handling}
\usage{
handle_warnings(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\value{
\code{list} of "result" which is the output of \code{expr} and "warnings"
which are any warnings that were generated during eval.
}
\description{
Utility to gracefully handle expected warnings
}
\examples{
f <- function() {
    warning("this is a warning")
    return(TRUE)
}
handle_warnings(f())
}
