% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{init_option}
\alias{init_option}
\title{Initialize an option}
\usage{
init_option(x, value)
}
\arguments{
\item{x}{character. Option to initialize}

\item{value}{value to set}
}
\value{
NULL invisibly
}
\description{
Set an option only if it does not exist yet.
}
\examples{
init_option("dummy_option", 2)
init_option("dummy_option", 5)
getOption("dummy_option") # still 2
}
