% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_flex.R
\name{lapply_flex}
\alias{lapply_flex}
\title{lapply_flex}
\usage{
lapply_flex(
  X,
  FUN,
  method = c("future", "biocparallel"),
  cores = NA,
  future.seed = TRUE,
  BPPARAM,
  fun = NULL,
  ...
)
}
\arguments{
\item{X}{list to use}

\item{FUN}{function to be performed}

\item{method}{character. Either \code{"future"} or \code{"biocparallel"}, determining
which parallelization framework to use.}

\item{cores}{deprecated}

\item{future.seed}{whether to set a seed when using \code{method = "future"}}

\item{BPPARAM}{a BiocParallel parameter class deciding how to perform
parallelized or sequential (default) evaluation. Used with
\code{method = "biocparallel"}}

\item{fun}{deprecated. Backwards compatibility for FUN}

\item{...}{other arguments to pass}
}
\value{
list
}
\description{
lapply_flex
}
\examples{
lapply_flex(list(x = 1, y = 2), FUN = log)

# suppress warnings
options("giotto.warn_sequential" = FALSE)

lapply_flex(list(x = 1, y = 2), FUN = log, method = "future")

lapply_flex(list(x = 1, y = 2), FUN = log, method = "biocparallel")
}
\keyword{internal}
