% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{list_input}
\alias{list_input}
\title{Coerce input into list of inputs (if not already)}
\usage{
list_input(x, dnames)
}
\arguments{
\item{x}{input}

\item{dnames}{character. default name(s) to apply IF resulting list is not named}
}
\description{
Coerce input into list of inputs (if not already)
}
