% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{local_seed}
\alias{local_seed}
\title{Set a seed local to a call}
\usage{
local_seed(seed)
}
\arguments{
\item{seed}{seed value to set}
}
\description{
When used inside a call, \code{local_seed()} first records the existing seed
then transiently sets a specific one. When the call exits, the recorded seed
is set again, so that the transient seed setting leaves no effects.\cr
Based on discussion from \url{https://support.bioconductor.org/p/110439/}
}
\examples{
f <- function() {
  local_seed(1234)
  r_val <- rnorm(1)
  return(r_val)
}

rnorm(1) # make sure a seed exists (not needed but handy for this example)

seed1 <- .Random.seed
x <- f()
seed2 <- .Random.seed
y <- rnorm(1)
seed3 <- .Random.seed
z <- f()

identical(seed1, seed2)
!identical(seed1, seed3)

identical(x, z)
!identical(x, y)
}
