% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_write}
\alias{log_write}
\title{Write to log file}
\usage{
log_write(file_conn = .log_conn(), x = "", collapse = " ", main = NULL)
}
\arguments{
\item{file_conn}{a file connection (Uses last created or generates a new one
if previous does not exist.)}

\item{x}{character vector. Content to write}

\item{collapse}{character. Collapse to use with \code{x}}

\item{main}{character. Title to assign log entry}
}
\value{
character
}
\description{
Write to log file
}
\examples{
log_write(x = "my content")

}
