% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_check.R
\name{new_github_ver_avail}
\alias{new_github_ver_avail}
\title{Check if a package has newer github version}
\usage{
new_github_ver_avail(url, current_ver = NULL)
}
\arguments{
\item{url}{character. url to the package to check's DESCRIPTION file.}

\item{current_ver}{character. Current version to check against}
}
\value{
character. Version number
}
\description{
This function works by downloading the DESCRIPTION file from a github
repo and then comparing the verison number with a provided value. If
the download fails or the github version is not newer than what is
installed, then it silently returns NULL. If the github version is newer,
the new version number will be returned
}
\examples{
url <- "https://raw.githubusercontent.com/giotto-suite/GiottoData/master/DESCRIPTION"
new_github_ver_avail(url, 0.2)
}
