% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{See \code{magrittr::\link[magrittr]{\%>\%}} for details.}
\usage{
lhs \%>\% rhs
}
\value{
pipe operator
}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\examples{
x <- "a"
x \%>\% print()
}
\keyword{internal}
