% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_call.R
\name{prev_call}
\alias{prev_call}
\alias{get_prev_call}
\alias{get_prev_fname}
\alias{get_args}
\alias{get_args_list}
\title{Previous calls}
\usage{
get_prev_call(toplevel = 1L)

get_prev_fname(toplevel = 3L)

get_args(toplevel = 2L, verbose = FALSE)

get_args_list(toplevel = 1L, keep = NULL, ...)
}
\arguments{
\item{toplevel}{integer. Relative stack where the function call was made.}

\item{verbose}{be verbose}

\item{keep}{character. When NULL, all params are captured. If not NULL,
specifies which params to capture.}

\item{\dots}{additional params to capture}
}
\value{
character
}
\description{
Functions to check the call stack and get aspects of previous
calls. Used in putting together object histories and error handling.
}
\section{Functions}{
\itemize{
\item \code{get_prev_call()}: Get previous call

\item \code{get_prev_fname()}: Get previous call function name

\item \code{get_args()}: Get previous call args as named character vector

\item \code{get_args_list()}: Get call args as named list

}}
\examples{
get_prev_call()

get_prev_fname(1)

# preserve params
foo <- function(a, b) get_args(toplevel = 1)
foo(a = 1, b = 2)

bar <- function() get_args() # default toplevel = 2
baz <- function(x, y) bar()
baz("a", "b")
a <- function(x = 1, y = 2, ...) {
    get_args_list(...)
}

a(z = 3, keep = "y")
}
