% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{print_list}
\alias{print_list}
\title{Pretty print formatting for lists and vectors}
\usage{
print_list(x, pre = "")
}
\arguments{
\item{x}{list of items to print. All entries must be named and have
\code{as.character()} methods}

\item{pre}{character. Optional characters to place at the head of each line}
}
\value{
lists and vectors
}
\description{
Pretty print formatting for lists and vectors
}
\examples{
print_list(list())
print_list(c())

testvec <- seq(3)
names(testvec) <- LETTERS[seq(3)]
print_list(testvec)

test <- list(
    name1 = "1",
    longername2 = "test_char",
    thirdname = factor("this will be converted with as.character()")
)
print_list(test)
print_list(test, pre = "* ")
}
