% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_utils.R
\name{py_active_env}
\alias{py_active_env}
\title{Active python environment}
\usage{
py_active_env()
}
\value{
boolean
}
\description{
Check for an active python environment without initialization.
If none initialized, \code{FALSE} is returned. If an initialized environment
is found, the env name based on \code{\link[reticulate:conda_list]{reticulate::conda_list()}} will be returned.
The following options will also be updated with information about the active
environment:
\itemize{
\item \code{"giotto.py_active_env"} \emph{(envname)}
\item \code{"giotto.py_active_ver"} \emph{(python version)}
\item \code{"giotto.py_active_path"} \emph{(path to python)}
}
}
\examples{
py_active_env()
}
