% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read.R
\name{read_colmatch}
\alias{read_colmatch}
\alias{fread_colmatch}
\title{Fread specific rows based on column matches}
\usage{
read_colmatch(file, col, sep = NULL, values_to_match, verbose = FALSE, ...)

fread_colmatch(...)
}
\arguments{
\item{file}{path to file to load}

\item{col}{name of col to match from}

\item{sep}{grep term to match as column delimiters within the file}

\item{values_to_match}{values in \code{col} to match given as a vector}

\item{verbose}{be verbose}

\item{...}{additional parameters to pass to \code{\link[arrow:open_delim_dataset]{arrow::open_delim_dataset()}}}
}
\value{
A tibble
}
\description{
Deprecated. Do not use
}
\section{Functions}{
\itemize{
\item \code{fread_colmatch()}: deprecated.

}}
\examples{
f <- file.path(tempdir(), "my_file.csv")
x <- data.frame(a = c("a", "b", "c"), b = 1:3, c = 5:7)
write.csv(x, f)
read_colmatch(f, col = "a", values_to_match = c("a", "c"))
unlink(f)

}
\keyword{internal}
