% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_utils.R
\name{scipy_sparse}
\alias{scipy_sparse}
\alias{to_scipy_sparse}
\alias{from_scipy_sparse}
\title{Convert between Matrix and scipy sparse representations}
\usage{
to_scipy_sparse(x, format = c("C", "R"), transpose = FALSE, ...)

from_scipy_sparse(x, format = c("C", "R"), transpose = FALSE, ...)
}
\arguments{
\item{x}{matrix-like object to convert}

\item{format}{character. Either "C" for Compressed Sparse Column or "R"
for Compressed Sparse Row matrix representations}

\item{transpose}{whether to transpose the matrix. default is FALSE}

\item{\dots}{additional params to pass to \verb{scipy.sparse.cs*_matrix()} or
\code{Matrix::sparseMatrix} depending on which output is desired.}
}
\value{
scipy or Matrix sparse representation depending on 'to' or 'from'
respectively
}
\description{
Convert between R \pkg{Matrix} sparse matrix representation and python scipy
sparse formats. Inspired by the implementation in \pkg{nmslibR}. Currently
compatible classes:\cr
\itemize{
\item \strong{\code{to_scipy_sparse()}}: \code{matrix}, \code{dgCMatrix}, \code{dgRMatrix}, \code{dgTMatrix}
\item \strong{\code{from_scipy_sparse()}}: \code{scipy.sparse._csr.csr_matrix},
\code{scipy.sparse._csc.csc_matrix}
}
}
\examples{
# example data
m <- matrix(data = 0L, nrow = 400, ncol = 300)
m[sample(400, 100), sample(300, 100)] <- runif(100, max = 10)
dgc <- Matrix::rsparsematrix(nrow = 400L, ncol = 300L, density = 0.1)

# some conversions
py_m_c <- to_scipy_sparse(m, format = "C")
py_m_r_t <- to_scipy_sparse(m, format = "R", transpose = TRUE)

py_dgc_c <- to_scipy_sparse(dgc, format = "C")
py_dgc_r <- to_scipy_sparse(dgc, format = "R")

dgc_revert <- from_scipy_sparse(py_dgc_c, format = "C")
dgr_revert <- from_scipy_sparse(py_dgc_c, format = "R")
identical(dgc, dgc_revert)
}
