% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{set_if}
\alias{set_if}
\alias{`\%null\%`,}
\alias{`\%none\%`,}
\alias{`\%na\%`}
\alias{\%null\%}
\alias{\%na\%}
\alias{\%none\%}
\title{Set if}
\usage{
x \%null\% y

x \%na\% y

x \%none\% y
}
\arguments{
\item{x}{object to test}

\item{y}{default value to set}
}
\value{
if \code{x} tests \code{TRUE}, \code{y} is returned
if \code{x} tests \code{FALSE} \code{x} is returned
}
\description{
Set a default value if a test is true.
}
\examples{
NULL \%null\% 1
2 \%null\% 1

"test" \%na\% 10
NA_character_ \%na\% "missing"

c() \%none\% 20
list() \%none\% "a"
logical(5) \%none\% 10

}
