% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{special_characters}
\alias{special_characters}
\alias{is_latex_output}
\alias{is_utf8_output}
\alias{box_chars}
\title{Special characters}
\usage{
is_latex_output()

is_utf8_output()

box_chars()
}
\description{
Framework and functions for printing with special characters within R packages.
}
\details{
Much inspiration taken from \pkg{fs} \href{https://rdrr.io/cran/fs/src/R/tree.R}{tree.R}
These are derived from: \href{https://github.com/r-lib/cli/blob/e9acc82b0d20fa5c64dd529400b622c0338374ed/R/tree.R#L111}{code}
}
\section{Functions}{
\itemize{
\item \code{is_latex_output()}: Determine if print is latex output

\item \code{is_utf8_output()}: Determine if system is using UTF-8 encoding

\item \code{box_chars()}: Helper function to print unicode box characters using escape codes.

}}
\keyword{internal}
