% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{standardize_param}
\alias{standardize_param}
\title{Standardize argument names}
\usage{
standardize_param(args_list, params)
}
\arguments{
\item{args_list}{a list of parameters to check}

\item{params}{character. Possible aliases for the parameter}
}
\value{
returns NULL if none are present. If only one is present, the value
of that param is returned
}
\description{
When multiple aliases are provided for a single argument, this function
looks through the provided list of parameters for up to 1 match and returns
the value of that matched param.
}
\examples{
f1 <- function(col = NULL, color = NULL, colour = NULL) {
  a <- get_args_list()
  col <- standardize_param(a, c("col", "color", "colour"))
  print(col)
}
f2 <- function(col = NA, color = NA, colour = NA) {
  a <- get_args_list()
  col <- standardize_param(a, c("col", "color", "colour"))
  print(col)
}
f3 <- function(...) {
  a <- list(...)
  col <- standardize_param(a, c("col", "color", "colour"))
  print(col)
}

f1(col = "green")
f2(color = "red")
f3(colour = "blue")
}
