% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{str_abbreviate}
\alias{str_abbreviate}
\title{Abbreviate a string}
\usage{
str_abbreviate(string, width = 60L, head = 15L, tail = 35L)
}
\arguments{
\item{string}{character. Input string}

\item{width}{numeric. Strings longer than this many characters will be
abbreviated}

\item{head}{numeric. Number of characters to include before abbreviated
section}

\item{tail}{numeric. Number of characters to include after abbreviated
section}
}
\value{
character
}
\description{
Abbreviates a string in the format of \code{head}[...]\code{tail}
when it exceeds the length specified by \code{width}. Useful for shortening
how filepaths are displayed.
}
\examples{
a <- "/short/file/path/"
b <- "/much/longer/foooooooooooooooooooo/baaaaaaaaaaaaaaaaaaaar/file/path/"
str_abbreviate(c(a,b))
str_abbreviate(c(a,b), width = 10, head = 3, tail = 3)
}
