% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{suite_packages}
\alias{suite_packages}
\title{Giotto Suite Packages}
\usage{
suite_packages(type = "core")
}
\arguments{
\item{type}{character. Type of module to return. Current choices are \code{"core"}
for the packages needed for Giotto to run, \code{"extensions"} for extending
packages such as \emph{GiottoData}, and \code{"all"} for all current modules to be
returned.}
}
\value{
character
}
\description{
Returns character vector of Giotto Suite's packages. Only the
core packages are returned by default.
}
\examples{
suite_packages()
suite_packages("core")
suite_packages("extensions")
suite_packages("all")
}
