% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{time_format}
\alias{time_format}
\title{Format time for printing}
\usage{
time_format(secs)
}
\arguments{
\item{secs}{numeric. seconds}
}
\value{
character
}
\description{
Format time for printing
}
\details{
Code from \code{\link[data.table]{timetaken}}
}
\examples{
time_format(90)

}
