% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_utils.R
\name{to_scipy_sparse}
\alias{to_scipy_sparse}
\title{Convert R sparse matrix to scipy sparse representation}
\usage{
to_scipy_sparse(x, format = c("C", "R"), transpose = FALSE, ...)
}
\arguments{
\item{x}{matrix-like object to convert}

\item{format}{character. Either "C" for Compressed Sparse Column or "R"
for Compressed Sparse Row matrix representations}

\item{transpose}{whether to transpose the matrix. default is FALSE}

\item{\dots}{additional params to pass to \verb{scipy.sparse.cs*_matrix()}}
}
\description{
Convert an R \pkg{Matrix} sparse matrix representation to python scipy
sparse format. Inspired by the implementation in \pkg{nmslibR}. Currently
works only for \code{matrix}, \code{dgCMatrix}, \code{dgRMatrix}, and \code{dgTMatrix}
}
