% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{vmsg}
\alias{vmsg}
\title{Verbose message handler}
\usage{
vmsg(
  ...,
  .v = NULL,
  .is_debug = FALSE,
  .vopt = getOption("giotto.verbose", TRUE)
)
}
\arguments{
\item{...}{additional strings and/or elements to pass to wrap_msg}

\item{.v}{verbose flag to pass. Will check options through .vopt if NULL (default).}

\item{.is_debug}{flag as a debug print (only prints when .v or .vopt is 'debug')}

\item{.vopt}{verbosity option to pull from}
}
\description{
Verbose message handler
}
