% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_helpers.R
\name{dt_set_row_order}
\alias{dt_set_row_order}
\title{Set specific data.table row order}
\usage{
dt_set_row_order(x, neworder)
}
\arguments{
\item{x}{data.table}

\item{neworder}{numerical vector to reorder rows}
}
\value{
A data.table
}
\description{
Set specific data.table row order
}
\examples{
x <- data.table::data.table(x = c("a", "b", "c"), y = 1:3, z = 5:7)
dt_set_row_order(x, neworder = c(1, 3, 2))
x

}
\concept{data.table}
