% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_helpers.R
\name{dt_to_matrix}
\alias{dt_to_matrix}
\title{dt_to_matrix}
\usage{
dt_to_matrix(x)
}
\arguments{
\item{x}{data.table object}
}
\value{
A matrix
}
\description{
converts data.table to matrix
}
\examples{
x = data.table::data.table(x = c("a","b","c"), y = 1:3, z = 5:7)
dt_to_matrix(x)

}
\concept{data.table}
