% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{getDistinctColors}
\usage{
getDistinctColors(n, seed = 1234)
}
\arguments{
\item{n}{number of colors wanted}

\item{seed}{seed to use when randomizing color order. Default is 1234}
}
\value{
character vector of hexadecimal distinct colors
}
\description{
Returns a number of distinct colors based on the RGB scale
}
\seealso{
Other basic color palette functions: 
\code{\link{getMonochromeColors}()},
\code{\link{getRainbowColors}()}
}
\concept{basic color palette functions}
