% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{getDistinctColors}
\usage{
getDistinctColors(n)
}
\arguments{
\item{n}{number of colors wanted}
}
\value{
character vector of hexadecimal distinct colors
}
\description{
Returns a number of distinct colors based on the RGB scale
}
\seealso{
Other basic color palette functions: 
\code{\link{getMonochromeColors}()},
\code{\link{getRainbowColors}()}
}
\concept{basic color palette functions}
