% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pals.R
\name{getRainbowColors}
\alias{getRainbowColors}
\title{getRainbowColors}
\usage{
getRainbowColors(n)
}
\arguments{
\item{n}{number of colors wanted}
}
\value{
character vector of hexadecimal rainbow colors
}
\description{
Returns a number of rainbow colors spaced around the spectrum.
Only 100 unique colors will be supplied after which they are recycled.
}
\seealso{
Other basic color palette functions: 
\code{\link{getDistinctColors}()},
\code{\link{getMonochromeColors}()}
}
\concept{basic color palette functions}
