% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_flex.R
\name{giotto_bpparam}
\alias{giotto_bpparam}
\title{Giotto Default BiocParallel Param}
\usage{
giotto_bpparam(BPPARAM = NULL)
}
\arguments{
\item{BPPARAM}{set a BiocParallel parameter class deciding how to perform
parallelized or (or sequential) evaluation}
}
\value{
If a \code{BiocParallelParam} is passed to \code{BPPARAM}, it will be cached
and invisibly returned. If \code{BPPARAM} is \code{NULL} then the cached param will be
invisibly returned (\code{SerialParam}) is default.
}
\description{
Get and set a cached BiocParallel BPPARAM setting. Cached settings can
either be retrieved using \code{giotto_bpparam()} or directly with
\code{getOption("giotto.bpparam")}
}
