% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_empty_char}
\alias{is_empty_char}
\title{Check if missing or empty character}
\usage{
is_empty_char(x)
}
\description{
Convenient function to test if a character input is NULL, missing,
or has a length of 0 (empty)
}
\keyword{internal}
