% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_flex.R
\name{lapply_flex}
\alias{lapply_flex}
\title{lapply_flex}
\usage{
lapply_flex(X, FUN, cores = NA, future.seed = TRUE, fun = NULL, ...)
}
\arguments{
\item{X}{list to use}

\item{FUN}{function to be performed}

\item{cores}{cores to use}

\item{future.seed}{whether to set a seed when using future_lapply}

\item{fun}{deprecated. Backwards compatibility for FUN}

\item{...}{other arguments to pass}
}
\value{
list
}
\description{
lapply_flex
}
\examples{
lapply_flex(list(x = 1, y = 2), FUN = log)

}
\keyword{internal}
