% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle_badge}
\alias{lifecycle_badge}
\title{lifecycle badge}
\usage{
lifecycle_badge(stage = "stable")
}
\arguments{
\item{stage}{character. 'stable', 'experimental', 'deprecated', or 'superseded'}
}
\description{
Generate markdown needed for lifecycle badges that help communicate stability
of functions in documentation when added in backticks as shown below.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ #' `r lifecycle::badge("experimental")`
 #' `r lifecycle::badge("deprecated")`
 #' `r lifecycle::badge("superseded")`
}\if{html}{\out{</div>}}
}
\seealso{
Other lifecycle: 
\code{\link{reexports}}
}
\concept{lifecycle}
