% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{melt_matrix}
\alias{melt_matrix}
\title{Melt a matrix}
\usage{
melt_matrix(data, value.name = "value")
}
\arguments{
\item{data}{\code{matrix} to melt}

\item{value.name}{name of variable used to store values (default = "value")}
}
\description{
Simple implementation of melt for matrices to produces results
similar to that of reshape2's now that it is deprecated. \cr
The params \code{varnames}, \code{na.rm} and, and \code{as.is} are not implemented.
}
\examples{
set.seed(9)
Data <- matrix(round(rnorm(12, 10, 4)), nrow = 4, ncol = 3)
melt_matrix(Data)
}
