% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_utils.R
\name{py_active_env}
\alias{py_active_env}
\title{Active python environment}
\usage{
py_active_env()
}
\description{
Check for an active python environment without initialization.
If none initialized, \code{FALSE} is returned. If an initialized environment
is found, the env name based on \code{\link[reticulate:conda-tools]{reticulate::conda_list()}} will be returned
}
