% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R, R/mixedsort.R, R/seed.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{deprecated}
\alias{is_present}
\alias{deprecate_soft}
\alias{deprecate_warn}
\alias{deprecate_stop}
\alias{mixedsort}
\alias{mixedorder}
\alias{withSeed}
\title{Objects exported from other packages}
\value{
argument marked as deprecated

boolean

NULL, invisibly.

NULL, invisibly.

NULL, invisibly.

A sorted vector

vector giving the sort order of the input elements
}
\seealso{
Other lifecycle: 
\code{\link{lifecycle_badge}()}
}
\concept{lifecycle}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gtools}{\code{\link[gtools:mixedsort]{mixedorder}}, \code{\link[gtools]{mixedsort}}}

  \item{lifecycle}{\code{\link[lifecycle]{deprecate_soft}}, \code{\link[lifecycle:deprecate_soft]{deprecate_stop}}, \code{\link[lifecycle:deprecate_soft]{deprecate_warn}}, \code{\link[lifecycle]{deprecated}}, \code{\link[lifecycle:deprecated]{is_present}}}

  \item{R.utils}{\code{\link[R.utils]{withSeed}}}
}}

