% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{str_convenience}
\alias{str_convenience}
\alias{str_vector}
\alias{str_bracket}
\alias{str_parenth}
\alias{str_double_quote}
\alias{str_quote}
\title{String convenience functions}
\usage{
str_vector(x, qchar = c("single", "double"))

str_bracket(x)

str_parenth(x)

str_double_quote(x)

str_quote(x)
}
\arguments{
\item{x}{string item(s) to format}

\item{qchar}{quote character to use. Either 'single' or "double"}
}
\description{
String convenience functions
}
\examples{
x <- "test"
cat(str_bracket(x), "\n")
cat(str_parenth(x), "\n")
cat(str_double_quote(x), "\n")
cat(str_quote(x), "\n")

vec <- c("item1", "item2", "item3")
cat(str_vector(vec), "\n")
cat(str_vector(vec, qchar = "double"))
}
