% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{suite_install}
\alias{suite_install}
\title{Giotto Suite GitHub Installation}
\usage{
suite_install(modules = suite_packages(), suite_deps = TRUE, ref = "main", ...)
}
\arguments{
\item{modules}{character. Which modules to install. Defaults to the core
packages needed for Giotto to run.}

\item{suite_deps}{logical. Whether to install any potential Giotto Suite
dependency modules}

\item{ref}{character. Currently one of "main", "R4.4.0", or "dev". These
determine which branches to install. See details.}

\item{\dots}{additional params to pass to \code{remotes::install_github()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Installation convenience function. Helps with Giotto Suite installation from
the GitHub website by ensuring modules are installed in the correct order.
Also automatically switches to the R4.4.0 version when user R version is
lower than 4.4.1.\cr

This utility will do a full reinstall of the specified Giotto Suite modules
but non-Suite dependencies will never prompt to be updated.
}
\section{ref \code{"main"}}{

Installs the main Giotto version. This version is expected to chase the
highest R version.
}

\section{ref \code{"R4.4.0"}}{

Mirrors the main branch, but locks the R version at 4.4.0
}

\section{ref \code{"dev"}}{

This version is ahead of the main version and also chases the latest R
version.
}

\examples{
if (FALSE) {
    # install core packages
    suite_install()
    suite_install("GiottoClass", ref = "dev")

    # install ONLY Giotto, ignoring module dependencies
    # (i.e. GiottoVisuals, GiottoClass, etc)
    suite_install("Giotto", suite_deps = FALSE)
}
}
