% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{vmsg}
\alias{vmsg}
\title{Verbose message handler}
\usage{
vmsg(
  ...,
  .v = NULL,
  .is_debug = FALSE,
  .vopt = getOption("giotto.verbose", TRUE)
)
}
\arguments{
\item{...}{additional strings and/or elements to pass to wrap_msg}

\item{.v}{verbose flag to pass. Will check options through .vopt if NULL (default).}

\item{.is_debug}{flag as a debug print (only prints when .v or .vopt is 'debug')}

\item{.vopt}{global verbosity option to pull from}
}
\description{
Verbose message handler
}
\examples{
# common usage (.v is logical)
vmsg('print me', .v = TRUE)
vmsg('dont print me', .v = FALSE)

# debug messages (.v == "debug")
# flag as a debug message using .is_debug = TRUE
vmsg('I am a debug message', .is_debug = TRUE, .v = TRUE) # no print
vmsg('I am a debug message', .is_debug = TRUE, .v = "debug") # prints

vmsg('print me', .v = 'debug') # also prints non-debug messages

# with global option
options("giotto.verbose" = TRUE)
vmsg("Print by default")

options("giotto.verbose" = FALSE)
vmsg("Do not print by default")
vmsg("Do not print by default", .v = TRUE) # function level input overrides global option
}
