% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with}
\alias{with}
\alias{gwith_options}
\alias{gwith_package}
\title{With utilities}
\usage{
gwith_options(new, code)

gwith_package(package, code, pos = 2L)
}
\arguments{
\item{new}{new option to set and its value}

\item{code}{code to execute with temporary settings}

\item{package}{character. The name of a package}

\item{pos}{integer. position in search to attach at. Default = 2}
}
\description{
Simple \emph{with} functions. Similar to or from \pkg{withr} implementations.
}
\section{Functions}{
\itemize{
\item \code{gwith_options()}: Eval with temporary option setting

\item \code{gwith_package()}: Eval with temporarily attached package

}}
\examples{
# options ###################################
gwith_options(list(gutils.temp = "found"), {
    print(getOption("gutils.temp", default = "not_found"))
})
getOption("gutils.temp", default = "not_found")

# packages ##################################
# temporarily attach at end of search (right before "base")
gwith_package("data.table", print(search()), pos = length(search()))
search()
}
