% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with}
\alias{with}
\alias{gwith_options}
\alias{gwith_package}
\alias{gwith_seed}
\title{With utilities}
\usage{
gwith_options(new, code)

gwith_package(package, code, pos = 2L)

gwith_seed(seed = 1234, code, ..., substitute = TRUE, envir = parent.frame())
}
\arguments{
\item{new}{new option to set and its value}

\item{code}{R code to execute with temporary settings}

\item{package}{character. The name of a package}

\item{pos}{integer. position in search to attach at. Default = 2}

\item{\dots}{additional params to pass. See details.}

\item{substitute}{If \code{\link[base:logical]{TRUE}}, argument \code{expr} is
   \code{\link[base]{substitute}()}:ed, otherwise not.}

\item{envir}{The \code{\link[base]{environment}} in which the expression should be evaluated.}
}
\description{
Simple \emph{with} functions. Similar to or from \pkg{withr} implementations.
}
\details{
\code{gwith_seed()} : \code{...} passes to \code{set.seed()}
}
\section{Functions}{
\itemize{
\item \code{gwith_options()}: Eval with temporary option setting

\item \code{gwith_package()}: Eval with temporarily attached package

\item \code{gwith_seed()}: Eval with temporary specifiable seed

}}
\examples{
# options ###################################
gwith_options(list(gutils.temp = "found"), {
    print(getOption("gutils.temp", default = "not_found"))
})
getOption("gutils.temp", default = "not_found")

# packages ##################################
# temporarily attach at end of search (right before "base")
gwith_package("data.table", print(search()), pos = length(search()))
search()

# seed  #####################################
start_seed <- .Random.seed

# identical generation
a <- gwith_seed(runif(10), seed = 1234)
b <- gwith_seed(runif(10), seed = 1234)
identical(a, b)

# does not alter pre-existing seed
end_seed <- .Random.seed
identical(start_seed, end_seed)
}
