% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{with_pbar}
\alias{with_pbar}
\title{Track progress while evaluating an R expression}
\usage{
with_pbar(
  expr,
  handlers = progressr::handlers(),
  cleanup = TRUE,
  delay_terminal = NULL,
  delay_stdout = NULL,
  delay_conditions = NULL,
  interrupts = getOption("progressr.interrupts", TRUE),
  interval = NULL,
  enable = NULL
)
}
\description{
Track progress while evaluating an R expression. This is a wrapper around
\code{\link[progressr:with_progress]{progressr::with_progress()}}. \pkg{progressr} is useful because it handles
progress reporting in parallel contexts. However, it is worth noting that
Giotto does not adhere to \pkg{progressr}'s guideline that \code{with_progress()}
should not be used within packages.
The default behavior of displaying progress is more convenient for end
users of packages with many functions. This implementation may shift over
time.
}
\examples{
a <- function(x = 10) {
    with_pbar({
        pb <- pbar(x)
        for (i in seq_len(x)) {
            Sys.sleep(0.1)
            pb()
        }
    })
}

a()
}
\seealso{
\code{\link[=pbar]{pbar()}}
}
