% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{wrap_txt}
\alias{wrap_txt}
\title{Wrap text}
\usage{
wrap_txt(
  ...,
  sep = " ",
  strWidth = 100,
  errWidth = FALSE,
  .prefix = " ",
  .initial = ""
)
}
\arguments{
\item{...}{additional params to pass}

\item{sep}{how to join elements of string (default is one space)}

\item{strWidth}{externally set wrapping width. (default value of 100
is not effected)}

\item{errWidth}{default = FALSE. Set strWidth to be compatible
with error printout}

\item{.prefix}{character. prefix for later lines}

\item{.initial}{character. prefix for first line}
}
\value{
character
}
\description{
Wrap text
}
\examples{
wrap_txt("A text")

}
